Move update/delete from TblVersionPath into TblVersion
